* keeper: support --wal-dir
For those who run Postgres with a WAL directory symlinked to a separate
disk (often for performance reasons) this commit adds a --wal-dir flag
to the keeper.

The symlink to a new WAL disk needs to be created as part of the
database initialisation step, and also whenever we clobber the database,
such as when taking a pg_basebackup (if pg_rewind is
disabled/unavailable). By passing this directory as an argument to
initdb and pg_basebackup, we can cover all the scenarios that the keeper
erases and recreates the data directory, ensuring the symlink is always
present.

* Support waldir in pitr command

For those who use different devices for wal and data directory, we
should support the interpolation of the wal directory location in the
pitr command as we do the data directory. This allows people to supply
commands like:

```
pg_basebackup --pgdata %d --waldir %w ...
```

And have both data directory and wal directory templated. pg_basebackup
and initdb are commands that support the concept of waldir, but this
could be useful for any arbitrary pitr command.

The commit also simplifies the implementation of the expand function to
feature less mutation and slice offset accesses.

* Also remove wal directory when destroying data

If a wal directory is supplied then it's important to clean-out our wals
along with our data directory. This commit renames the existing
RemoveAll method to RemoveAllIfInitialized, which better represents what
it does, and has it rely on a RemoveAll method that cleans-up both data
directory and wal.

Co-authored-by: Harry Maclean <harrymaclean@gocardless.com>